+++ src/EventListener/GraphQL/AccessControlListener.php
@@ -16,6 +16,7 @@ use Ynlo\GraphQLBundle\Events\GraphQLFieldEvent;
 use Ynlo\GraphQLBundle\Events\GraphQLMutationEvent;
 use Ynlo\GraphQLBundle\Exception\Controlled\ForbiddenError;
 use Ynlo\GraphQLBundle\Security\Authorization\AccessControlChecker;
+use Ynlo\GraphQLBundle\Util\TypeUtil;
 
 class AccessControlListener implements EventSubscriberInterface
 {
@@ -56,7 +57,7 @@ class AccessControlListener implements EventSubscriberInterface
 
     public function preReadField(GraphQLFieldEvent $event)
     {
-        //check firstly if the user have rights to read the object
+        //check firstly if the user have rights to read the operation
         $node = $event->getContext()->getNode();
         if ($node && $this->accessControlChecker->isControlled($node)
             && !$this->accessControlChecker->isGranted($node, $event->getContext()->getRoot())
@@ -64,6 +65,21 @@ class AccessControlListener implements EventSubscriberInterface
             $event->stopPropagation();
             $event->setValue(null);
             throw new ForbiddenError($this->accessControlChecker->getMessage($node));
+        }
+
+        //check if user have rights to read the object
+        if (\is_object($event->getContext()->getRoot())) {
+            $concreteType = TypeUtil::resolveObjectType($event->getContext()->getEndpoint(), $event->getContext()->getRoot());
+            if ($concreteType) {
+                $objectDefinition = $event->getContext()->getEndpoint()->getType($concreteType);
+                if ($this->accessControlChecker->isControlled($objectDefinition)
+                    && !$this->accessControlChecker->isGranted($objectDefinition, $event->getContext()->getRoot())
+                ) {
+                    $event->stopPropagation();
+                    $event->setValue(null);
+                    throw new ForbiddenError($this->accessControlChecker->getMessage($objectDefinition));
+                }
+            }
         }
 
         //check then if the user have rights to read the field