+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 v1.2.2 - (Unreleased)
 ----
  * Added support to use collection filters for interfaces (polymorphic nodes)
+ * Added support to set filters type with modifiers `[]` or `!` 
   
 v1.2.1 - 2018-07-16
 ----