+++ CHANGELOG.md
@@ -5,6 +5,7 @@ v1.2.2 - (Unreleased)
  * Added support to use page based pagination
  * Fixed issue checking permissions on specific objects
  * Fixed fields from interface use same endpoints as interface definition
+ * Fixed issue in lists not ordering correctly parent fields
   
 v1.2.1 - 2018-07-16
 ----