+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-v1.2.2 - (Unreleased)
+v1.2.2 - 2018-07-25
 ----
  * Added support to use collection filters for interfaces (polymorphic nodes)
  * Added support to set filters type with modifiers `[]` or `!`